feature/UI
1.계속 출력되던 보스 HP 출력문 삭제(UI반영됨)
2.HP버그 확인 중 총알에 맞았을 때 받는 데미지 등 실시간 UI반영안됨 확인
3. HP버그 해결 1차시도(자료형 변경)

+++ Assets/GameManager.cs
@@ -6,7 +6,7 @@ public class GameManager : MonoBehaviour
     public static GameManager gameManager;
 
     public float playTime = 0f;
-    public float playerHp;
+    public int playerHp;
 
     [Header("Player Damage")]
     public int playerBaseDamage = 1;
